refactor(sdk): split request and response types into two files
