+++ sdk/src/rpc/types.ts
@@ -1,5 +1,8 @@
 import { Address, Hex } from 'viem';
 
+export * from './request-types';
+export * from './response-types';
+
 /** Type alias for Request ID (uint64) */
 export type RequestID = number;
 
@@ -339,680 +342,3 @@ export enum RPCMethod {
     Message = 'message'
 }
 
-/**
- * Represents the parameters for the 'auth_challenge' RPC method.
- */
-export interface AuthChallengeRPCParams {
-    challengeMessage: string;
-}
-
-/**
- * Represents the parameters for the 'auth_verify' RPC method.
- */
-export interface AuthVerifyRPCParams {
-    address: Address;
-    jwtToken: string;
-    sessionKey: Address;
-    success: boolean;
-}
-
-/**
- * Represents the parameters for the 'error' RPC method.
- */
-export interface ErrorRPCParams {
-    error: string;
-}
-
-/**
- * Represents the parameters for the 'get_config' RPC method.
- */
-export interface GetConfigRPCParams {
-    broker_address: Address;
-    networks: {
-        name: string;
-        chain_id: number;
-        custody_address: Address;
-        adjudicator_address: Address;
-    }[];
-}
-
-/**
- * Represents the parameters for the 'get_ledger_balances' RPC method.
- */
-export interface GetLedgerBalancesRPCParams {
-    asset: string;
-    amount: string;
-}
-
-/**
- * Represents the parameters for the 'get_ledger_entries' RPC method.
- */
-export interface GetLedgerEntriesRPCParams {
-    id: number;
-    account_id: string;
-    account_type: string;
-    asset: string;
-    participant: Address;
-    credit: string;
-    debit: string;
-    created_at: string;
-}
-
-/**
- * Represents the parameters for the 'create_app_session' RPC method.
- */
-export interface CreateAppSessionRPCParams {
-    app_session_id: Hex;
-    version: number;
-    status: ChannelStatus;
-}
-
-/**
- * Represents the parameters for the 'submit_state' RPC method.
- */
-export interface SubmitStateRPCParams {
-    app_session_id: Hex;
-    version: number;
-    status: ChannelStatus;
-}
-
-/**
- * Represents the parameters for the 'close_app_session' RPC method.
- */
-export interface CloseAppSessionRPCParams {
-    app_session_id: Hex;
-    version: number;
-    status: ChannelStatus;
-}
-
-/**
- * Represents the parameters for the 'get_app_definition' RPC method.
- */
-export interface GetAppDefinitionRPCParams extends AppDefinition { }
-
-/**
- * Represents the parameters for the 'get_app_sessions' RPC method.
- */
-export interface GetAppSessionsRPCParams {
-    app_session_id: Hex;
-    status: ChannelStatus;
-    participants: Address[];
-    protocol: string;
-    challenge: number;
-    weights: number[];
-    quorum: number;
-    version: number;
-    nonce: number;
-    created_at: string;
-    updated_at: string;
-}
-
-/**
- * Represents the parameters for the 'resize_channel' RPC method.
- */
-export interface ResizeChannelRPCParams {
-    channel_id: Hex;
-    state_data: string;
-    intent: number;
-    version: number;
-    allocations: {
-        destination: Address;
-        token: Address;
-        amount: string;
-    }[];
-    state_hash: string;
-    server_signature: {
-        v: string;
-        r: string;
-        s: string;
-    };
-}
-
-/**
- * Represents the parameters for the 'close_channel' RPC method.
- */
-export interface CloseChannelRPCParams {
-    channel_id: Hex;
-    intent: number;
-    version: number;
-    state_data: string;
-    allocations: {
-        destination: Address;
-        token: Address;
-        amount: string;
-    }[];
-    state_hash: string;
-    server_signature: {
-        v: string;
-        r: string;
-        s: string;
-    };
-}
-
-/**
- * Represents the parameters for the 'get_channels' RPC method.
- */
-export interface GetChannelsRPCParams {
-    channel_id: Hex;
-    participant: Address;
-    status: ChannelStatus;
-    token: Address;
-    wallet: Address;
-    amount: string;
-    chain_id: number;
-    adjudicator: Address;
-    challenge: number;
-    nonce: number;
-    version: number;
-    created_at: string;
-    updated_at: string;
-}
-
-/**
- * Represents the parameters for the 'get_rpc_history' RPC method.
- */
-export interface GetRPCHistoryRPCParams {
-    id: number;
-    sender: Address;
-    req_id: number;
-    method: string;
-    params: string;
-    timestamp: number;
-    req_sig: Hex[];
-    res_sig: Hex[];
-    response: string;
-}
-
-/**
- * Represents the parameters for the 'get_assets' RPC method.
- */
-export interface GetAssetsRPCParams {
-    token: Address;
-    chain_id: number;
-    symbol: string;
-    decimals: number;
-}
-
-/**
- * Represents a generic RPC message structure that includes common fields.
- * This interface is extended by specific RPC request and response types.
- */
-interface GenericRPCMessage {
-    requestId: RequestID;
-    timestamp?: Timestamp;
-    signatures?: Hex[];
-}
-
-/**
- * Represents the request structure for the 'auth_challenge' RPC method.
- */
-export interface AuthChallengeRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.AuthChallenge;
-    params: AuthChallengeRPCRequestParams;
-}
-
-/**
- * Represents the response structure for the 'auth_challenge'
- */
-export interface AuthChallengeRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.AuthChallenge;
-    params: AuthChallengeRPCParams;
-}
-
-/**
- * Represents the request parameters for the 'auth_challenge' RPC method.
- */
-export interface AuthChallengeRPCRequestParams {
-    /** The challenge message to be signed by the client for authentication. */
-    challengeMessage: string;
-}
-
-/**
- * Represents the request parameters for the 'auth_verify' RPC method.
- */
-export interface AuthVerifyRPCRequestParams {
-    /** The Ethereum address of the client attempting to authenticate. */
-    address: Address;
-    /** JSON Web Token for authentication, if provided. */
-    jwtToken: string;
-    /** The session key address associated with the authentication attempt. */
-    sessionKey: Address;
-    /** Indicates whether the authentication attempt was successful. */
-    success: boolean;
-}
-
-/**
- * Represents the request structure for the 'auth_verify' RPC method.
- */
-export interface AuthVerifyRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.AuthVerify;
-    params: AuthVerifyRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'get_config' RPC method.
- */
-export interface GetConfigRPCRequestParams {
-    /** The Ethereum address of the broker. */
-    broker_address: Address;
-    /** List of supported networks and their configurations. */
-    networks: {
-        /** The name of the network (e.g., "Ethereum", "Polygon"). */
-        name: string;
-        /** The chain ID of the network. */
-        chain_id: number;
-        /** The custody contract address for the network. */
-        custody_address: Address;
-        /** The adjudicator contract address for the network. */
-        adjudicator_address: Address;
-    }[];
-}
-
-/**
- * Represents the request structure for the 'get_config' RPC method.
- */
-export interface GetConfigRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.GetConfig;
-    params: GetConfigRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'get_ledger_balances' RPC method.
- */
-export interface GetLedgerBalancesRPCRequestParams {
-    /** Optional participant address to filter balances. If not provided, uses the authenticated wallet address. */
-    participant?: Address;
-    /** Optional account ID to filter balances. If provided, overrides the participant address. */
-    account_id?: string;
-}
-
-/**
- * Represents the request structure for the 'get_ledger_balances' RPC method.
- */
-export interface GetLedgerBalancesRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.GetLedgerBalances;
-    params: GetLedgerBalancesRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'get_ledger_entries' RPC method.
- */
-export interface GetLedgerEntriesRPCRequestParams {
-    /** Optional account ID to filter ledger entries. */
-    account_id?: string;
-    /** Optional asset symbol to filter ledger entries. */
-    asset?: string;
-    /** Optional wallet address to filter ledger entries. If provided, overrides the authenticated wallet. */
-    wallet?: Address;
-}
-
-/**
- * Represents the request structure for the 'get_ledger_entries' RPC method.
- */
-export interface GetLedgerEntriesRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.GetLedgerEntries;
-    params: GetLedgerEntriesRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'create_app_session' RPC method.
- */
-export interface CreateAppSessionRPCRequestParams {
-    /** The detailed definition of the application being created, including protocol, participants, weights, and quorum. */
-    definition: AppDefinition;
-    /** The initial allocation distribution among participants. Each participant must have sufficient balance for their allocation. */
-    allocations: AppSessionAllocation[];
-}
-
-/**
- * Represents the request structure for the 'create_app_session' RPC method.
- */
-export interface CreateAppSessionRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.CreateAppSession;
-    params: CreateAppSessionRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'submit_state' RPC method.
- */
-export interface SubmitStateRPCRequestParams {
-    /** The unique identifier of the application session to update. */
-    app_session_id: Hex;
-    /** The new allocation distribution among participants. Must include all participants and maintain total balance. */
-    allocations: AppSessionAllocation[];
-}
-
-/**
- * Represents the request structure for the 'submit_state' RPC method.
- */
-export interface SubmitStateRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.SubmitState;
-    params: SubmitStateRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'close_app_session' RPC method.
- */
-export interface CloseAppSessionRPCRequestParams {
-    /** The unique identifier of the application session to close. */
-    app_session_id: Hex;
-    /** The final allocation distribution among participants upon closing. Must include all participants and maintain total balance. */
-    allocations: AppSessionAllocation[];
-}
-
-/**
- * Represents the request structure for the 'close_app_session' RPC method.
- */
-export interface CloseAppSessionRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.CloseAppSession;
-    params: CloseAppSessionRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'get_app_definition' RPC method.
- */
-export interface GetAppDefinitionRPCRequestParams {
-    /** The unique identifier of the application session to retrieve. */
-    app_session_id: Hex;
-}
-
-/**
- * Represents the request structure for the 'get_app_definition' RPC method.
- */
-export interface GetAppDefinitionRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.GetAppDefinition;
-    params: GetAppDefinitionRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'get_app_sessions' RPC method.
- */
-export interface GetAppSessionsRPCRequestParams {
-    /** Optional participant address to filter application sessions. If not provided, returns all sessions. */
-    participant?: Address;
-    /** Optional status to filter application sessions (e.g., "open", "closed"). If not provided, returns sessions of all statuses. */
-    status?: ChannelStatus;
-}
-
-/**
- * Represents the request structure for the 'get_app_sessions' RPC method.
- */
-export interface GetAppSessionsRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.GetAppSessions;
-    params: GetAppSessionsRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'resize_channel' RPC method.
- */
-export interface ResizeChannelRPCRequestParams {
-    /** The unique identifier of the channel to resize. */
-    channel_id: Hex;
-    /** Optional amount to resize the channel by (can be positive or negative). Must be provided if allocate_amount is not. */
-    resize_amount?: bigint;
-    /** Optional amount to allocate from the unified balance to the channel. Must be provided if resize_amount is not. */
-    allocate_amount?: bigint;
-    /** The address where the resized funds will be sent. */
-    funds_destination: Address;
-}
-
-/**
- * Represents the request structure for the 'resize_channel' RPC method.
- */
-export interface ResizeChannelRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.ResizeChannel;
-    params: ResizeChannelRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'close_channel' RPC method.
- */
-export interface CloseChannelRPCRequestParams {
-    /** The unique identifier of the channel to close. */
-    channel_id: Hex;
-    /** The address where the channel funds will be sent upon closing. */
-    funds_destination: Address;
-}
-
-/**
- * Represents the request structure for the 'close_channel' RPC method.
- */
-export interface CloseChannelRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.CloseChannel;
-    params: CloseChannelRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'get_channels' RPC method.
- */
-export interface GetChannelsRPCRequestParams {
-    /** Optional participant address to filter channels. If not provided, returns all channels. */
-    participant?: Address;
-    /** Optional status to filter channels (e.g., "open", "closed"). If not provided, returns channels of all statuses. */
-    status?: ChannelStatus;
-}
-
-/**
- * Represents the request structure for the 'get_channels' RPC method.
- */
-export interface GetChannelsRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.GetChannels;
-    params: GetChannelsRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'get_rpc_history' RPC method.
- */
-export interface GetRPCHistoryRPCRequestParams {
-    /** The participant address to retrieve RPC history for. Must be the authenticated wallet address. */
-    participant: Address;
-}
-
-/**
- * Represents the request structure for the 'get_rpc_history' RPC method.
- */
-export interface GetRPCHistoryRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.GetRPCHistory;
-    params: GetRPCHistoryRPCRequestParams;
-}
-
-/**
- * Represents the request parameters for the 'get_assets' RPC method.
- */
-export interface GetAssetsRPCRequestParams {
-    /** Optional chain ID to filter assets by network. If not provided, returns assets from all networks. */
-    chain_id?: number;
-}
-
-/**
- * Represents the request structure for the 'get_assets' RPC method.
- */
-export interface GetAssetsRPCRequest extends GenericRPCMessage {
-    method: RPCMethod.GetAssets;
-    params: GetAssetsRPCRequestParams;
-}
-
-/**
- * Union type for all possible RPC request types.
- * This allows for type-safe handling of different request structures.
- */
-export type RPCRequest =
-    | AuthChallengeRPCRequest
-    | AuthVerifyRPCRequest
-    | GetConfigRPCRequest
-    | GetLedgerBalancesRPCRequest
-    | GetLedgerEntriesRPCRequest
-    | CreateAppSessionRPCRequest
-    | SubmitStateRPCRequest
-    | CloseAppSessionRPCRequest
-    | GetAppDefinitionRPCRequest
-    | GetAppSessionsRPCRequest
-    | ResizeChannelRPCRequest
-    | CloseChannelRPCRequest
-    | GetChannelsRPCRequest
-    | GetRPCHistoryRPCRequest
-    | GetAssetsRPCRequest;
-
-/**
- * Represents the response structure for an error response.
- */
-export interface ErrorRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.Error;
-    params: ErrorRPCParams;
-}
-
-/**
- * Represents the response structure for the 'get_config' RPC method.
- */
-export interface GetConfigRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.GetConfig;
-    params: GetConfigRPCParams;
-}
-
-/**
- * Represents the response structure for the 'get_ledger_balances' RPC method.
- */
-export interface GetLedgerBalancesRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.GetLedgerBalances;
-    params: GetLedgerBalancesRPCParams[];
-}
-
-/**
- * Represents the response structure for the 'get_ledger_entries' RPC method.
- */
-export interface GetLedgerEntriesRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.GetLedgerEntries;
-    params: GetLedgerEntriesRPCParams[];
-}
-
-/**
- * Represents the response structure for the 'create_app_session' RPC method.
- */
-export interface CreateAppSessionRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.CreateAppSession;
-    params: CreateAppSessionRPCParams;
-}
-
-/**
- * Represents the response structure for the 'submit_state' RPC method.
- */
-export interface SubmitStateRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.SubmitState;
-    params: SubmitStateRPCParams;
-}
-
-/**
- * Represents the response structure for the 'close_app_session' RPC method.
- */
-export interface CloseAppSessionRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.CloseAppSession;
-    params: CloseAppSessionRPCParams;
-}
-
-/**
- * Represents the response structure for the 'get_app_definition' RPC method.
- */
-export interface GetAppDefinitionRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.GetAppDefinition;
-    params: GetAppDefinitionRPCParams;
-}
-
-/**
- * Represents the response structure for the 'get_app_sessions' RPC method.
- */
-export interface GetAppSessionsRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.GetAppSessions;
-    params: GetAppSessionsRPCParams[];
-}
-
-/**
- * Represents the response structure for the 'resize_channel' RPC method.
- */
-export interface ResizeChannelRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.ResizeChannel;
-    params: ResizeChannelRPCParams;
-}
-
-/**
- * Represents the response structure for the 'close_channel' RPC method.
- */
-export interface CloseChannelRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.CloseChannel;
-    params: CloseChannelRPCParams;
-}
-
-/**
- * Represents the response structure for the 'get_channels' RPC method.
- */
-export interface GetChannelsRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.GetChannels;
-    params: GetChannelsRPCParams[];
-}
-
-/**
- * Represents the response structure for the 'get_rpc_history' RPC method.
- */
-export interface GetRPCHistoryRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.GetRPCHistory;
-    params: GetRPCHistoryRPCParams[];
-}
-
-/**
- * Represents the response structure for the 'get_assets' RPC method.
- */
-export interface GetAssetsRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.GetAssets;
-    params: GetAssetsRPCParams[];
-}
-
-/**
- * Represents the response structure for the 'auth_verify' RPC method.
- */
-export interface AuthVerifyRPCResponse extends GenericRPCMessage {
-    method: RPCMethod.AuthVerify;
-    params: AuthVerifyRPCParams;
-}
-
-/**
- * Union type for all possible RPC response types.
- * This allows for type-safe handling of different response structures.
- */
-export type RPCResponse =
-    | AuthChallengeRPCResponse
-    | AuthVerifyRPCResponse
-    | ErrorRPCResponse
-    | GetConfigRPCResponse
-    | GetLedgerBalancesRPCResponse
-    | GetLedgerEntriesRPCResponse
-    | CreateAppSessionRPCResponse
-    | SubmitStateRPCResponse
-    | CloseAppSessionRPCResponse
-    | GetAppDefinitionRPCResponse
-    | GetAppSessionsRPCResponse
-    | ResizeChannelRPCResponse
-    | CloseChannelRPCResponse
-    | GetChannelsRPCResponse
-    | GetRPCHistoryRPCResponse
-    | GetAssetsRPCResponse;
-
-/**
- * Maps RPC methods to their corresponding parameter types.
- */
-export type RPCParamsByMethod = {
-    [RPCMethod.AuthChallenge]: AuthChallengeRPCParams;
-    [RPCMethod.AuthVerify]: AuthVerifyRPCParams;
-    [RPCMethod.Error]: ErrorRPCParams;
-    [RPCMethod.GetConfig]: GetConfigRPCParams;
-    [RPCMethod.GetLedgerBalances]: GetLedgerBalancesRPCParams[];
-    [RPCMethod.GetLedgerEntries]: GetLedgerEntriesRPCParams[];
-    [RPCMethod.CreateAppSession]: CreateAppSessionRPCParams;
-    [RPCMethod.SubmitState]: SubmitStateRPCParams;
-    [RPCMethod.CloseAppSession]: CloseAppSessionRPCParams;
-    [RPCMethod.GetAppDefinition]: GetAppDefinitionRPCParams;
-    [RPCMethod.GetAppSessions]: GetAppSessionsRPCParams[];
-    [RPCMethod.ResizeChannel]: ResizeChannelRPCParams;
-    [RPCMethod.CloseChannel]: CloseChannelRPCParams;
-    [RPCMethod.GetChannels]: GetChannelsRPCParams[];
-    [RPCMethod.GetRPCHistory]: GetRPCHistoryRPCParams[];
-    [RPCMethod.GetAssets]: GetAssetsRPCParams[];
-};
